chore: renaming sync hook
