+++ frontend/dashboard/src/app/[community]/dashboard/Client.tsx
@@ -2,14 +2,14 @@
 
 import { DSB_ROUTE } from '~/const/route'
 import SideMenu from '~/containers/thread/DashboardThread/SideMenu'
-import useDSBRouteTabSync, { isRouteOf } from '~/hooks/useDSBRouteTabSync'
+import useSyncDSBRoute2Tab, { isRouteOf } from '~/hooks/useSyncDSBRoute2Tab'
 import CommunityDigest from '~/widgets/CommunityDigest'
 import useSalon from './salon'
 
 const Layout = ({ children }) => {
   const s = useSalon()
 
-  useDSBRouteTabSync({
+  useSyncDSBRoute2Tab({
     tab: 'curTab',
     defaultTab: DSB_ROUTE.OVERVIEW,
     segmentIndex: 2,
